web/requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
django-chartjs>=1.5.0
django-rest-swagger